wrote ver1 of algo and converted it to js classes
diverged from using an array of day objects to an array that holds
timeslots that act like linked lists.

loosing data on the day such as percent used but that's ok for now.

timeslots holds events and tasks. tasks will just be added like events
after an appropriate start time is found for them.